requirements.txt
python-telegram-bot==20.6
pdfminer.six==20230729
python-docx==0.9.0
